Sane microscopy image labeling in Julia

Sane microscopy image labeling in Julia

This is collection of simple labeling functions specifically suited for quickly embedding labels in microscopy images. I'm hoping to cover common use-cases like adding timestamps, scalebars, transitions, and extra image details.

MicroscopyLabels.timestamp! — Method.

timestamp!(img; units)

Writes the timestamp in the upper left corner of a multidimensional image img with the units units. Images are required to have at least the x, y, and time axes labeled, i.e. img should be or contain an AxisArray object with xyt axes. If additional axes are present, the timestamp will be written to all of them.

Simple example

using AxisArrays
using MicroscopyLabels

tmp = AxisArray(zeros(200, 200, 5), Axis{:x}(1:200), Axis{:y}(1:200), Axis{:time}(1:5))
timestamp!(tmp)

Unitful example

We can also add units to the time axis and have these be embedded in the image

using Unitful: s, minute

tmp = AxisArray(zeros(200, 200, 5), Axis{:x}(1:200), Axis{:y}(1:200),
                Axis{:time}(0s:45.0s:180s))
# convert the units to minutes
timestamp!(tmp, units=minute)

source